# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Meme Token API",
    "version": "1.0.0",
    "description": "API for minting and managing meme tokens on Solana blockchain",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/meme/token-balance/{tokenAccountAddress}": {
      "get": {
        "tags": ["Mint Details"],
        "summary": "Get token account balance",
        "description": "Retrieves the balance of a specific token account",
        "operationId": "getTokenBalance",
        "parameters": [
          {
            "name": "tokenAccountAddress",
            "in": "path",
            "required": true,
            "description": "The public key address of the token account",
            "schema": {
              "type": "string",
              "example": "HN7cABqLq46Es1jh92dQQisAq662SmxELLLsHHe4YWrH"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenBalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token account address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Token account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/meme/meme-token-distribution/{memeId}": {
      "get": {
        "tags": ["Mint Details"],
        "summary": "Get meme token distribution",
        "description": "Retrieves the distribution details of a meme token including minter and vault balances",
        "operationId": "getMemeTokenDistribution",
        "parameters": [
          {
            "name": "memeId",
            "in": "path",
            "required": true,
            "description": "The meme ID to query",
            "schema": {
              "type": "string",
              "example": "meme_abc123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Distribution retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemeTokenDistributionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid meme ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Meme token not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/meme/meme-token-balances": {
      "post": {
        "tags": ["Mint Details"],
        "summary": "Get multiple meme token balances",
        "description": "Retrieves balance information for multiple meme tokens",
        "operationId": "getAllMemeTokenBalances",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemeTokenBalancesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Balances retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemeTokenBalancesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/meme/meme-token-recent/{memeId}": {
      "get": {
        "tags": ["Mint Details"],
        "summary": "Get recent mint distribution",
        "description": "Retrieves formatted distribution details for a recently minted token",
        "operationId": "getRecentMintDistribution",
        "parameters": [
          {
            "name": "memeId",
            "in": "path",
            "required": true,
            "description": "The meme ID to query",
            "schema": {
              "type": "string",
              "example": "meme_abc123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Distribution retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentMintDistributionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid meme ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Meme token not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/meme/all-minted-tokens": {
      "get": {
        "tags": ["Mint Details"],
        "summary": "Get all minted tokens",
        "description": "Retrieves a list of all meme tokens that have been minted across all users",
        "operationId": "getAllMintedTokens",
        "responses": {
          "200": {
            "description": "All minted tokens retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllMintedTokensResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/amm/pool-info/{tokenMint}": {
      "get": {
        "tags": ["AMM Pool"],
        "summary": "Get pool information",
        "description": "Retrieves detailed information about an AMM pool including reserves, LP supply, and pool addresses",
        "operationId": "getPoolInfo",
        "parameters": [
          {
            "name": "tokenMint",
            "in": "path",
            "required": true,
            "description": "The token mint address of the pool",
            "schema": {
              "type": "string",
              "example": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pool information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoolInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid token mint address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pool not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TokenBalanceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "balance": {
                "type": "string",
                "description": "Raw balance amount",
                "example": "1000000000000000000"
              },
              "balanceNumber": {
                "type": "number",
                "description": "Balance as a number",
                "example": 1000000000000000000
              },
              "decimals": {
                "type": "integer",
                "description": "Number of decimals",
                "example": 9
              },
              "mint": {
                "type": "object",
                "description": "Token mint public key object"
              }
            },
            "required": ["balance", "balanceNumber", "decimals", "mint"]
          }
        },
        "required": ["success", "data"]
      },
      "MemeTokenDistributionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "mint": {
                "type": "object",
                "description": "Token mint public key"
              },
              "totalSupply": {
                "type": "string",
                "example": "1000000000000000000"
              },
              "minterBalance": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "address": {
                    "type": "object",
                    "description": "Minter's token account address"
                  },
                  "balance": {
                    "type": "string",
                    "example": "20000000000000000"
                  },
                  "balanceNumber": {
                    "type": "number",
                    "example": 20000000000000000
                  },
                  "percentage": {
                    "type": "number",
                    "example": 2
                  }
                }
              },
              "vaultBalance": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "address": {
                    "type": "object",
                    "description": "Vault's token account address"
                  },
                  "balance": {
                    "type": "string",
                    "example": "980000000000000000"
                  },
                  "balanceNumber": {
                    "type": "number",
                    "example": 980000000000000000
                  },
                  "percentage": {
                    "type": "number",
                    "example": 98
                  }
                }
              },
              "distributionSummary": {
                "type": "object",
                "properties": {
                  "totalDistributed": {
                    "type": "number",
                    "example": 1000000000000000000
                  },
                  "minterShare": {
                    "type": "number",
                    "example": 2
                  },
                  "vaultShare": {
                    "type": "number",
                    "example": 98
                  },
                  "isCorrectDistribution": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "required": ["success", "data"]
      },
      "MemeTokenBalancesRequest": {
        "type": "object",
        "properties": {
          "memeIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["meme_abc123", "meme_def456"]
          }
        },
        "required": ["memeIds"]
      },
      "MemeTokenBalancesResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "memeId": {
                  "type": "string",
                  "example": "meme_abc123"
                },
                "memeIdHex": {
                  "type": "string",
                  "example": "6d656d655f616263313233"
                },
                "distribution": {
                  "type": "object",
                  "description": "Distribution details (same structure as MemeTokenDistributionResponse)"
                }
              }
            }
          }
        },
        "required": ["success", "data"]
      },
      "RecentMintDistributionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "mint": {
                "type": "string",
                "description": "Token mint address",
                "example": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
              },
              "totalSupply": {
                "type": "string",
                "example": "1000000000.0"
              },
              "minter": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "HN7cABqLq46Es1jh92dQQisAq662SmxELLLsHHe4YWrH"
                  },
                  "balance": {
                    "type": "string",
                    "example": "20000000.0"
                  },
                  "percentage": {
                    "type": "string",
                    "example": "2.0000%"
                  }
                }
              },
              "vault": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "J3dxNj7nDRRqRRXuEMynDG57DkZK4jYRuv3Garmb1i99"
                  },
                  "balance": {
                    "type": "string",
                    "example": "980000000.0"
                  },
                  "percentage": {
                    "type": "string",
                    "example": "98.0000%"
                  }
                }
              },
              "summary": {
                "type": "object",
                "properties": {
                  "totalDistributed": {
                    "type": "string",
                    "example": "1000000000.0"
                  },
                  "isCorrect": {
                    "type": "boolean",
                    "example": true
                  },
                  "minterShare": {
                    "type": "string",
                    "example": "2.0000%"
                  },
                  "vaultShare": {
                    "type": "string",
                    "example": "98.0000%"
                  }
                }
              }
            }
          }
        },
        "required": ["success", "data"]
      },
      "AllMintedTokensResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "All minted tokens retrieved successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "count": {
                "type": "integer",
                "description": "Total number of minted tokens",
                "example": 5
              },
              "tokens": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MintedToken"
                }
              }
            },
            "required": ["count", "tokens"]
          }
        },
        "required": ["success", "message", "data"]
      },
      "MintedToken": {
        "type": "object",
        "properties": {
          "memeId": {
            "type": "string",
            "description": "Human-readable meme ID",
            "example": "meme_abc123"
          },
          "memeIdHex": {
            "type": "string",
            "description": "Hexadecimal representation of meme ID",
            "example": "6d656d655f616263313233"
          },
          "mint": {
            "type": "string",
            "description": "Token mint address",
            "example": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
          },
          "minter": {
            "type": "string",
            "description": "Address of the account that minted the token",
            "example": "5ZWj7a1f8tWkjBESHKgrLmXshuXxqeY9SYcfbshpAqPG"
          },
          "createdAt": {
            "type": "string",
            "description": "Unix timestamp when the token was created",
            "example": "1640000000"
          },
          "isInitialized": {
            "type": "boolean",
            "description": "Whether the token state is initialized",
            "example": true
          }
        },
        "required": ["memeId", "memeIdHex", "mint", "minter", "createdAt", "isInitialized"]
      },
...
      "PoolInfoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "poolPda": {
                "type": "string",
                "description": "Pool PDA address",
                "example": "8sLbNZoA1cfnvMJLPfp98ZLAnFSYCFApfJKMbiXNLwxj"
              },
              "tokenMint": {
                "type": "string",
                "description": "Token mint address",
                "example": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
              },
              "lpMint": {
                "type": "string",
                "description": "LP token mint address",
                "example": "CKfatsPMUf8SkiURsDXs7eK6GWb4Jsd6UDbs7twMCWxo"
              },
              "solVault": {
                "type": "string",
                "description": "SOL vault address",
                "example": "J3dxNj7nDRRqRRXuEMynDG57DkZK4jYRuv3Garmb1i99"
              },
              "tokenVault": {
                "type": "string",
                "description": "Token vault address",
                "example": "HN7cABqLq46Es1jh92dQQisAq662SmxELLLsHHe4YWrH"
              },
              "solReserve": {
                "type": "string",
                "description": "SOL reserve in lamports",
                "example": "1500000000"
              },
              "tokenReserve": {
                "type": "string",
                "description": "Token reserve amount",
                "example": "1000000"
              },
              "lpSupply": {
                "type": "string",
                "description": "Total LP token supply",
                "example": "1224744"
              },
              "isInitialized": {
                "type": "boolean",
                "description": "Whether the pool is initialized",
                "example": true
              }
            }
          }
        },
        "required": ["success", "data"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          },
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid input provided"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "INVALID_INPUT"
          }
        },
        "required": ["success"]
      }
    }
  }
}